Split ChatInput runtime hooks
